refactor(profiling): split 'superdesk-profiling' module.
`superdesk-profiling` becomes:
```
.
├── controllers
│   ├── ProfilingController.js
│   └── index.js
├── styles
│   └── profiling.less
├── views
│   └── profiling.html
└── index.js

3 directories, 5 files
```
