+++ scripts/superdesk-profiling/controllers/index.js
@@ -1,0 +1,1 @@
+export { ProfilingController } from './ProfilingController';
